refactoring-#21 : Entity 구체화
