+++ src/main/java/com/web/userchat/model/MessageType.java
@@ -1,7 +1,7 @@
 package com.web.userchat.model;
 
 public enum MessageType {
-    CHAT,
-    SYSTEM,
-    SUBSCRIBE
+    CHAT, // 일반 채팅 메세지
+    ENTER, // 채팅방 입장
+    LEAVE // 채팅방 퇴장
 }
