--- src/main/java/com/web/userchat/repository/ChatMessageRepository.java
@@ -1,13 +1,0 @@
-package com.web.userchat.repository;
-
-import com.web.userchat.model.ChatMessage;
-import com.web.userchat.model.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-
-@Repository
-public interface ChatMessageRepository extends JpaRepository<ChatMessage, Long> {
-    List<ChatMessage> findByChatRoom(ChatRoom chatRoom);
-}
