--- src/main/java/com/web/userchat/repository/ChatRoomRepository.java
@@ -1,8 +1,0 @@
-package com.web.userchat.repository;
-
-import com.web.userchat.model.ChatRoom;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ChatRoomRepository extends JpaRepository<ChatRoom, String> {
-
-}
