--- src/main/java/com/web/userchat/repository/UserRepository.java
@@ -1,16 +1,0 @@
-package com.web.userchat.repository;
-
-import com.web.userchat.model.User;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.stereotype.Repository;
-
-import java.util.List;
-import java.util.Optional;
-
-@Repository
-public interface UserRepository extends JpaRepository<User, Long> {
-    Optional<User> findByEmail(String email);
-    List<User> findByUsernameContaining(String query);
-
-
-}
